requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
django==3.2; python_version >= '3.6'
djangorestframework==3.14.0; python_version >= '3.6'
pytz==2023.3.post1
sqlparse==0.4.4; python_version >= '3.5'
